Implement seveal changes in environment.
changed tmux-file-selector to use just plain fzf as we have updated
FZF_DEFAULT_COMMAND to use ripgrep now which is a lot faster than our
old version.

Several changes to vimrc to begin making vim a litte bit easier to use
in context of note taking. This includes settings for fzf and vimwiki.